Merge async query

+++ webui/src/components/pages/MyWorkspace/Table.jsx
@@ -1,89 +1,92 @@
-import { Table, Modal, Spin, Result,  } from "antd";
+import { Table, Modal, Spin, Result, Popover,  } from "antd";
 import { useContext, useEffect, useState,} from "react";
 import { DeleteOutlined, FileTextOutlined, LoadingOutlined } from "@ant-design/icons";
 import { useNavigate } from "react-router-dom";
 import { UserContext } from "../../pages/AppContent/AppContext";
-import { uploadDocument, deleteDocument, getAllDocsMetaData } from "../../../services/Api";
+import { uploadDocument, deleteDocument, getAllDocsMetaData, getInitialData } from "../../../services/Api";
 import DragFile from "./DragFile";
-import useFileStore from "./store";
+import useFileStore from "./fileStore";
 import GenericModal from "../../common/Modal/GenericModal"
-import { useQuery, useQueryClient } from "@tanstack/react-query";
-import { CheckOutlined } from "@mui/icons-material";
+import { useQuery } from "@tanstack/react-query";
+import { CheckOutlined, ErrorOutline } from "@mui/icons-material";
 
+export const LOADING = "loading";
+export const DONE = "done";
+export const ERROR = "error";
 
 function FileTable() {
   const navigate = useNavigate();
-  const [prefetched, setPrefetched] = useState(false)
   const [pdfFile, setPdfFile] = useState(null);
   const [fileMetaData, setFileMetaData] = useState(null);
 
   const { user, token } = useContext(UserContext);
   const [open, setOpen] = useState(false);
-  const [loading, setLoading] = useState(false);
-  const { files, addFileToStore, removeFileFromStore, setAllDocs } = useFileStore();
+  const [status, setStatus] = useState(false);
+  const { files, addFileToStore, removeFileFromStore, updateFileStatusInStore, setAllFiles } = useFileStore();
 
   /**
-   * {
-   *  file1: {metadata: null, pdfFile: null, loading: false },
-   *  file2: {metadata: null, pdfFile: null, loading: false }
-   * }
+   * [
+   *  file1: {name: file1, size: null, dateModified: null,  status: loading/done/error },
+   * ]
    */
 
-  const queryClient = useQueryClient()
-
-  const {data} = useQuery({
-      queryKey:["docs"], 
-      queryFn: () => getInitialData(user.email), 
-      enabled: user!=null,
-      refetchOnWindowFocus: false
-    },
-  )
-
-  // Fetch initial data
-  // ======================================================
-
-  const getInitialData = async (user_id) => {
-    console.log("get initial data.")
+// Fetch initial data
+// ======================================================
+  const getUserAllDocsMetaData = async (user_id, token) => {
+    if (!user_id || !token){return []};
     let initialDataResponse = await getAllDocsMetaData(user_id, token);
     if (initialDataResponse.status!==200 && initialDataResponse.status!==204){
-      console.log("An error occured while trying to fetch initial data.");
       return []
     }
     let docs = initialDataResponse.data ? initialDataResponse.data.docs_metadata : null
     if (docs == null){return [];}
-    
     docs = docs.map((d) => ({
       name: d.document_id,
       size: `${Math.round(d.document_size / 1024)} KB`,
-      dateModified: d.creation_time,
-    }))
+      dateModified: new Date(d.creation_time).toLocaleDateString(),
+      status: DONE
+    }));
+    setAllFiles(docs);
     return docs;
   }
 
-  const prefetch = async () => {
-    // if (!prefetched){
-      await queryClient.prefetchQuery({queryKey:['docs'], queryFn: () => getInitialData(user.email)});
-      setPrefetched(true);
-    // }
-  }
-  prefetch()
-
-  useEffect(()=>{
-    if (data){
-      setAllDocs(data)
-    }
-  },[data])
-  
+  const {data} = useQuery({
+      queryKey:["docs"], 
+      queryFn: () => getUserAllDocsMetaData(user.email, token), 
+      enabled: (token!==undefined && token!=null),
+      refetchOnWindowFocus: false,
+    },
+  )
 
   // Helpers and configs
   // ======================================================
+  const reset = () => {
+    setStatus(false);
+    setPdfFile(null);
+    setFileMetaData(null);
+  }
+
   const onCancel = () => {
-    if (!loading){
+    if (status !== LOADING){
       setOpen(false);
-      setLoading(false);
-      setPdfFile(null);
-      setFileMetaData(null);
+      reset();
     }    
+  }
+
+  const onRemove = () => {
+    console.log("remove.")
+    setPdfFile(null);
+    setFileMetaData(null);
+  }
+
+  const validateAscii = (name) => {
+    let pattern = /^[\x00-\x7F]+$/;
+    return pattern.test(name);
+  }
+
+  const validateIsUnique = (name) => {
+    const matchingFiles = files.filter(f => f.name === name)
+    return (matchingFiles.length === 0)
   }
 
   const columns = [
@@ -108,14 +111,20 @@ function FileTable() {
     {
       title: "Progress",
       key: "progress",
-      render: (record) => {
-        return (
-          <>
-            {loading ? <Spin indicator={<LoadingOutlined style={{fontSize:24}} />} /> : <CheckOutlined /> }
-          </>
-        )
+      render: (record) => {    
+        if (!record.status || record.status === LOADING) {
+          return <Spin indicator={<LoadingOutlined style={{ fontSize: 24 }} />} />;
+        } else if (record.status === DONE) {
+          return <CheckOutlined style={{ color: "green" }} />;
+        } else {
+          return (
+            <Popover content={"Error while proccessing file. Please delete and try again."} trigger="hover">
+              <ErrorOutline style={{ color: "red" }}/>
+            </Popover>   
+          )
+        }
       }
-    },
+    },    
     {
       title: "Action",
       key: "action",
@@ -125,7 +134,7 @@ function FileTable() {
             <DeleteOutlined onClick={() => onDeleteFile(record)} />
             <FileTextOutlined
               style={{ color: "black", marginLeft: 10 }}
-              onClick={() => navigate("/doc-view/" + record.name)}
+              onClick={() => onDisplayFile(record)}
             />
           </>
         );
@@ -133,31 +142,54 @@ function FileTable() {
     },
   ];
 
+  // display file
+  // ======================================================
+  const onDisplayFile = (record) =>{
+    if (record.status !== DONE ) { return ; }
+    navigate("/doc-view/" + record.name)
+  }
+
+
   // Delete a File
   // ======================================================
   const onDeleteFile = (record) => {
-    Modal.confirm({
-      title: "Are you sure you want to delete this file?",
-      okText: "yes",
-      cancelText: "No",
-      okType: "danger",
-      onOk: () => {
-        deleteDocument(user.email, record.name, token); // backend
-        removeFileFromStore(record.name);
-      },
-    });
-  };
+    if (record.status === LOADING ) { return ; }
+    if (record.status === DONE){
+      Modal.confirm({
+        title: "Are you sure you want to delete this file?",
+        okText: "yes",
+        cancelText: "No",
+        okType: "danger",
+        onOk: () => {
+          deleteDocument(user.email, record.name, token); // backend
+          removeFileFromStore(record.name);
+        },
+      });
+    } else {
+      Modal.confirm({
+        title: "Remove record from table",
+        okText: "yes",
+        cancelText: "No",
+        okType: "danger",
+        onOk: () => {
+          removeFileFromStore(record.name);
+        },
+      });
+    }
+  }
+
 
   // Upload a File
   // ======================================================
 
   const onUploadFile = async () => {
-    setLoading(true);
+    setStatus(true);
     try {
       const newFile = {
         name: fileMetaData.name.split(".")[0],
         size: `${Math.round(fileMetaData.size / 1024)} KB`,
         dateModified: new Date().toLocaleDateString(),
+        status: LOADING
       };
   
       let filePayload = {
@@ -168,64 +200,79 @@ function FileTable() {
         "pdf_encoding": pdfFile
       }
 
-      console.log("token" +token)
       let uploadDocResponse = await uploadDocument(filePayload, token); // backend
-    
 
     // check if error occured when communicating with the backend
     if (uploadDocResponse.status!==200 && uploadDocResponse.status!==204){
-      alert("An error occured while trying uploading the file. Please try again.");
+      updateFileStatusInStore(newFile.name, ERROR);
       return; // finally will exectue of course.
     }
 
+    newFile.status = DONE;
     // Otherwise, it was successfull. Add the file to the table.
     addFileToStore(newFile);
 
     }
     finally{
       setOpen(false);
-      setPdfFile(null);
-      setFileMetaData(null);
-      setLoading(false);
+      reset();
     }
   }
 
   // Async upload a File
   // ======================================================
-
+  // file1: {name: file1, size: null, dateModified: null,  status: false }
   const onAsyncUploadFile = async () => {
-    setLoading(true);
+    if (!fileMetaData || ! pdfFile) {return ; }
+    setStatus(true);
+    // Make a deep copy so state can be used to upload additional files.
+    const pdfFileCopy = JSON.parse(JSON.stringify(pdfFile));
+    const fileMetaDataCopy = JSON.parse(JSON.stringify(fileMetaData));
     try {
       const newFile = {
-        name: fileMetaData.name.split(".")[0],
-        size: `${Math.round(fileMetaData.size / 1024)} KB`,
+        name: fileMetaDataCopy.name.split(".")[0],
+        size: `${Math.round(fileMetaDataCopy.size / 1024)} KB`,
         dateModified: new Date().toLocaleDateString(),
+        status: LOADING
       };
   
+      const name = newFile.name;
       let filePayload = {
         "document_metadata": {
             "user_id": user.email,
-            "document_id": newFile.name
+            "document_id": name
         },
-        "pdf_encoding": pdfFile
+        "pdf_encoding": pdfFileCopy
       }
-    addFileToStore(newFile);
 
-    uploadDocument(filePayload)
-      .then((uploadDocResponse) => {
-        // check if error occured when communicating with the backend
-        if (uploadDocResponse.status!==200 && uploadDocResponse.status!==204){
-          alert("An error occured while trying uploading the file. Please try again.");
-          return; // finally will exectue of course.
-        }
-        // Otherwise, it was successfull. Add the file to the table.
-        setLoading(false);
-      })
+      let valid = validateAscii(name);
+      if (!valid){
+        alert("Document name must be in ascii letters.") ;reset(); return;
+      }
+      valid = validateIsUnique(name)
+      if (!valid){
+        alert("Document: " + name + " is already exists") ;reset(); return;
+      }
+      
+      addFileToStore(newFile);
+      reset();
+      uploadDocument(filePayload, token)
+        .then((uploadDocResponse) => {
+          // check if error occured when communicating with the backend
+          if (uploadDocResponse.status!==200 && uploadDocResponse.status!==204){
+            updateFileStatusInStore(newFile.name, ERROR);
+            return; // finally will exectue of course.
+          }
+          // update the of this praticular file to done.
+          updateFileStatusInStore(newFile.name, DONE);
+        })
+        .catch( (error) => {
+          updateFileStatusInStore(newFile.name, ERROR);
+          }
+        )
     }
     finally{
       setOpen(false);
-      setPdfFile(null);
-      setFileMetaData(null);
     }
   }
 
@@ -234,10 +281,11 @@ function FileTable() {
     <div>
       <DragFile 
         onCancel={onCancel}
-        onSubmit={onUploadFile}
+        onRemove={onRemove}
+        onSubmit={onAsyncUploadFile}
         setFile={setPdfFile}
         setFileMetaData={setFileMetaData}
-        loading={loading}
+        loading={status}
       />
     </div>
 
@@ -253,12 +301,14 @@ function FileTable() {
       <GenericModal
             open={open} 
             setOpen={setOpen}
-            loading={loading}
-            setLoading={setLoading} 
-            onCancel={null}
+            loading={status}
+            setStatus={setStatus} 
+            onCancel={onCancel}
             modalButtonText="upload"
             modalTitle={"Upload File"}
-            modalContent={content}            
+            modalContent={content}
+            buttonType={"primary"}
+            buttonSize={"large"}            
       />
       <Table
         columns={columns}
